Move timestream module into submodules
