+++ awswrangler/timestream/__init__.py
@@ -1,0 +1,22 @@
+"""Amazon Timestream Module."""
+
+
+from awswrangler.timestream._create import create_database, create_table
+from awswrangler.timestream._delete import delete_database, delete_table
+from awswrangler.timestream._list import list_databases, list_tables
+from awswrangler.timestream._read import query
+from awswrangler.timestream._write import batch_load, batch_load_from_files, wait_batch_load_task, write
+
+__all__ = [
+    "create_database",
+    "create_table",
+    "delete_database",
+    "delete_table",
+    "list_databases",
+    "list_tables",
+    "query",
+    "write",
+    "batch_load",
+    "batch_load_from_files",
+    "wait_batch_load_task",
+]
