+++ awswrangler/timestream/_write.py
@@ -4,7 +4,7 @@ import itertools
 import logging
 import time
 from datetime import datetime
-from typing import TYPE_CHECKING, Any, Dict, Iterator, List, Literal, Optional, Union, cast, overload
+from typing import TYPE_CHECKING, Any, Dict, List, Optional, Union, cast
 
 import boto3
 from botocore.config import Config
@@ -18,7 +18,6 @@ from awswrangler.distributed.ray import ray_get
 from awswrangler.typing import TimestreamBatchLoadReportS3Configuration
 
 if TYPE_CHECKING:
-    from mypy_boto3_timestream_query.type_defs import PaginatorConfigTypeDef, QueryResponseTypeDef, RowTypeDef
     from mypy_boto3_timestream_write.client import TimestreamWriteClient
 
 _BATCH_LOAD_FINAL_STATES: List[str] = ["SUCCEEDED", "FAILED", "PROGRESS_STOPPED", "PENDING_RESUME"]
@@ -196,103 +195,6 @@ def _write_df(
         itertools.repeat(dimensions_cols),
         batches,
     )
-
-
-def _cast_value(value: str, dtype: str) -> Any:  # pylint: disable=too-many-branches,too-many-return-statements
-    if dtype == "VARCHAR":
-        return value
-    if dtype in ("INTEGER", "BIGINT"):
-        return int(value)
-    if dtype == "DOUBLE":
-        return float(value)
-    if dtype == "BOOLEAN":
-        return value.lower() == "true"
-    if dtype == "TIMESTAMP":
-        return datetime.strptime(value[:-3], "%Y-%m-%d %H:%M:%S.%f")
-    if dtype == "DATE":
-        return datetime.strptime(value, "%Y-%m-%d").date()
-    if dtype == "TIME":
-        return datetime.strptime(value[:-3], "%H:%M:%S.%f").time()
-    if dtype == "ARRAY":
-        return str(value)
-    raise ValueError(f"Not supported Amazon Timestream type: {dtype}")
-
-
-def _process_row(schema: List[Dict[str, str]], row: "RowTypeDef") -> List[Any]:
-    row_processed: List[Any] = []
-    for col_schema, col in zip(schema, row["Data"]):
-        if col.get("NullValue", False):
-            row_processed.append(None)
-        elif "ScalarValue" in col:
-            row_processed.append(_cast_value(value=col["ScalarValue"], dtype=col_schema["type"]))
-        elif "ArrayValue" in col:
-            row_processed.append(_cast_value(value=col["ArrayValue"], dtype="ARRAY"))  # type: ignore[arg-type]
-        else:
-            raise ValueError(
-                f"Query with non ScalarType/ArrayColumnInfo/NullValue for column {col_schema['name']}. "
-                f"Expected {col_schema['type']} instead of {col}"
-            )
-    return row_processed
-
-
-def _rows_to_df(
-    rows: List[List[Any]], schema: List[Dict[str, str]], df_metadata: Optional[Dict[str, str]] = None
-) -> pd.DataFrame:
-    df = pd.DataFrame(data=rows, columns=[c["name"] for c in schema])
-    if df_metadata:
-        try:
-            df.attrs = df_metadata
-        except AttributeError as ex:
-            # Modin does not support attribute assignment
-            _logger.error(ex)
-    for col in schema:
-        if col["type"] == "VARCHAR":
-            df[col["name"]] = df[col["name"]].astype("string")
-    return df
-
-
-def _process_schema(page: "QueryResponseTypeDef") -> List[Dict[str, str]]:
-    schema: List[Dict[str, str]] = []
-    for col in page["ColumnInfo"]:
-        if "ScalarType" in col["Type"]:
-            schema.append({"name": col["Name"], "type": col["Type"]["ScalarType"]})
-        elif "ArrayColumnInfo" in col["Type"]:
-            schema.append({"name": col["Name"], "type": col["Type"]["ArrayColumnInfo"]})
-        else:
-            raise ValueError(f"Query with non ScalarType or ArrayColumnInfo for column {col['Name']}: {col['Type']}")
-    return schema
-
-
-def _paginate_query(
-    sql: str,
-    chunked: bool,
-    pagination_config: Optional["PaginatorConfigTypeDef"],
-    boto3_session: Optional[boto3.Session] = None,
-) -> Iterator[pd.DataFrame]:
-    client = _utils.client(
-        service_name="timestream-query",
-        session=boto3_session,
-        botocore_config=Config(read_timeout=60, retries={"max_attempts": 10}),
-    )
-    paginator = client.get_paginator("query")
-    rows: List[List[Any]] = []
-    schema: List[Dict[str, str]] = []
-    page_iterator = paginator.paginate(QueryString=sql, PaginationConfig=pagination_config or {})
-    for page in page_iterator:
-        if not schema:
-            schema = _process_schema(page=page)
-            _logger.debug("schema: %s", schema)
-        for row in page["Rows"]:
-            rows.append(_process_row(schema=schema, row=row))
-        if len(rows) > 0:
-            df_metadata = {}
-            if chunked:
-                if "NextToken" in page:
-                    df_metadata["NextToken"] = page["NextToken"]
-                df_metadata["QueryId"] = page["QueryId"]
-
-            yield _rows_to_df(rows, schema, df_metadata)
-        rows = []
 
 
 @_utils.validate_distributed_kwargs(
@@ -759,363 +661,3 @@ def batch_load_from_files(
         timestream_batch_load_wait_polling_delay=timestream_batch_load_wait_polling_delay,
         boto3_session=boto3_session,
     )
-
-
-@overload
-def query(
-    sql: str,
-    chunked: Literal[False] = ...,
-    pagination_config: Optional[Dict[str, Any]] = ...,
-    boto3_session: Optional[boto3.Session] = ...,
-) -> pd.DataFrame:
-    ...
-
-
-@overload
-def query(
-    sql: str,
-    chunked: Literal[True],
-    pagination_config: Optional[Dict[str, Any]] = ...,
-    boto3_session: Optional[boto3.Session] = ...,
-) -> Iterator[pd.DataFrame]:
-    ...
-
-
-@overload
-def query(
-    sql: str,
-    chunked: bool,
-    pagination_config: Optional[Dict[str, Any]] = ...,
-    boto3_session: Optional[boto3.Session] = ...,
-) -> Union[pd.DataFrame, Iterator[pd.DataFrame]]:
-    ...
-
-
-def query(
-    sql: str,
-    chunked: bool = False,
-    pagination_config: Optional[Dict[str, Any]] = None,
-    boto3_session: Optional[boto3.Session] = None,
-) -> Union[pd.DataFrame, Iterator[pd.DataFrame]]:
-    """Run a query and retrieve the result as a Pandas DataFrame.
-
-    Parameters
-    ----------
-    sql: str
-        SQL query.
-    chunked: bool
-        If True returns DataFrame iterator, and a single DataFrame otherwise. False by default.
-    pagination_config: Dict[str, Any], optional
-        Pagination configuration dictionary of a form {'MaxItems': 10, 'PageSize': 10, 'StartingToken': '...'}
-    boto3_session : boto3.Session(), optional
-        Boto3 Session. The default boto3 Session will be used if boto3_session receive None.
-
-    Returns
-    -------
-    Union[pd.DataFrame, Iterator[pd.DataFrame]]
-        Pandas DataFrame https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.DataFrame.html
-
-    Examples
-    --------
-    Run a query and return the result as a Pandas DataFrame or an iterable.
-
-    >>> import awswrangler as wr
-    >>> df = wr.timestream.query('SELECT * FROM "sampleDB"."sampleTable" ORDER BY time DESC LIMIT 10')
-
-    """
-    result_iterator = _paginate_query(sql, chunked, cast("PaginatorConfigTypeDef", pagination_config), boto3_session)
-    if chunked:
-        return result_iterator
-
-    # Prepending an empty DataFrame ensures returning an empty DataFrame if result_iterator is empty
-    results = list(result_iterator)
-    if len(results) > 0:
-        # Modin's concat() can not concatenate empty data frames
-        return pd.concat(results, ignore_index=True)
-    return pd.DataFrame()
-
-
-def create_database(
-    database: str,
-    kms_key_id: Optional[str] = None,
-    tags: Optional[Dict[str, str]] = None,
-    boto3_session: Optional[boto3.Session] = None,
-) -> str:
-    """Create a new Timestream database.
-
-    Note
-    ----
-    If the KMS key is not specified, the database will be encrypted with a
-    Timestream managed KMS key located in your account.
-
-    Parameters
-    ----------
-    database: str
-        Database name.
-    kms_key_id: Optional[str]
-        The KMS key for the database. If the KMS key is not specified,
-        the database will be encrypted with a Timestream managed KMS key located in your account.
-    tags: Optional[Dict[str, str]]
-        Key/Value dict to put on the database.
-        Tags enable you to categorize databases and/or tables, for example,
-        by purpose, owner, or environment.
-        e.g. {"foo": "boo", "bar": "xoo"})
-    boto3_session : boto3.Session(), optional
-        Boto3 Session. The default boto3 Session will be used if boto3_session receive None.
-
-    Returns
-    -------
-    str
-        The Amazon Resource Name that uniquely identifies this database. (ARN)
-
-    Examples
-    --------
-    Creating a database.
-
-    >>> import awswrangler as wr
-    >>> arn = wr.timestream.create_database("MyDatabase")
-
-    """
-    _logger.info("Creating Timestream database %s", database)
-    client = _utils.client(service_name="timestream-write", session=boto3_session)
-    args: Dict[str, Any] = {"DatabaseName": database}
-    if kms_key_id is not None:
-        args["KmsKeyId"] = kms_key_id
-    if tags is not None:
-        args["Tags"] = [{"Key": k, "Value": v} for k, v in tags.items()]
-    response = client.create_database(**args)
-    return response["Database"]["Arn"]
-
-
-def delete_database(
-    database: str,
-    boto3_session: Optional[boto3.Session] = None,
-) -> None:
-    """Delete a given Timestream database. This is an irreversible operation.
-
-    After a database is deleted, the time series data from its tables cannot be recovered.
-
-    All tables in the database must be deleted first, or a ValidationException error will be thrown.
-
-    Due to the nature of distributed retries,
-    the operation can return either success or a ResourceNotFoundException.
-    Clients should consider them equivalent.
-
-    Parameters
-    ----------
-    database: str
-        Database name.
-    boto3_session : boto3.Session(), optional
-        Boto3 Session. The default boto3 Session will be used if boto3_session receive None.
-
-    Returns
-    -------
-    None
-        None.
-
-    Examples
-    --------
-    Deleting a database
-
-    >>> import awswrangler as wr
-    >>> arn = wr.timestream.delete_database("MyDatabase")
-
-    """
-    _logger.info("Deleting Timestream database %s", database)
-    client = _utils.client(service_name="timestream-write", session=boto3_session)
-    client.delete_database(DatabaseName=database)
-
-
-def create_table(
-    database: str,
-    table: str,
-    memory_retention_hours: int,
-    magnetic_retention_days: int,
-    tags: Optional[Dict[str, str]] = None,
-    timestream_additional_kwargs: Optional[Dict[str, Any]] = None,
-    boto3_session: Optional[boto3.Session] = None,
-) -> str:
-    """Create a new Timestream database.
-
-    Note
-    ----
-    If the KMS key is not specified, the database will be encrypted with a
-    Timestream managed KMS key located in your account.
-
-    Parameters
-    ----------
-    database: str
-        Database name.
-    table: str
-        Table name.
-    memory_retention_hours: int
-        The duration for which data must be stored in the memory store.
-    magnetic_retention_days: int
-        The duration for which data must be stored in the magnetic store.
-    tags: Optional[Dict[str, str]]
-        Key/Value dict to put on the table.
-        Tags enable you to categorize databases and/or tables, for example,
-        by purpose, owner, or environment.
-        e.g. {"foo": "boo", "bar": "xoo"})
-    timestream_additional_kwargs : Optional[Dict[str, Any]]
-        Forwarded to botocore requests.
-        e.g. timestream_additional_kwargs={'MagneticStoreWriteProperties': {'EnableMagneticStoreWrites': True}}
-    boto3_session : boto3.Session(), optional
-        Boto3 Session. The default boto3 Session will be used if boto3_session receive None.
-
-    Returns
-    -------
-    str
-        The Amazon Resource Name that uniquely identifies this database. (ARN)
-
-    Examples
-    --------
-    Creating a table.
-
-    >>> import awswrangler as wr
-    >>> arn = wr.timestream.create_table(
-    ...     database="MyDatabase",
-    ...     table="MyTable",
-    ...     memory_retention_hours=3,
-    ...     magnetic_retention_days=7
-    ... )
-
-    """
-    _logger.info("Creating Timestream table %s in database %s", table, database)
-    client = _utils.client(service_name="timestream-write", session=boto3_session)
-    timestream_additional_kwargs = {} if timestream_additional_kwargs is None else timestream_additional_kwargs
-    args: Dict[str, Any] = {
-        "DatabaseName": database,
-        "TableName": table,
-        "RetentionProperties": {
-            "MemoryStoreRetentionPeriodInHours": memory_retention_hours,
-            "MagneticStoreRetentionPeriodInDays": magnetic_retention_days,
-        },
-        **timestream_additional_kwargs,
-    }
-    if tags is not None:
-        args["Tags"] = [{"Key": k, "Value": v} for k, v in tags.items()]
-    response = client.create_table(**args)
-    return response["Table"]["Arn"]
-
-
-def delete_table(
-    database: str,
-    table: str,
-    boto3_session: Optional[boto3.Session] = None,
-) -> None:
-    """Delete a given Timestream table.
-
-    This is an irreversible operation.
-
-    After a Timestream database table is deleted, the time series data stored in the table cannot be recovered.
-
-    Due to the nature of distributed retries,
-    the operation can return either success or a ResourceNotFoundException.
-    Clients should consider them equivalent.
-
-    Parameters
-    ----------
-    database: str
-        Database name.
-    table: str
-        Table name.
-    boto3_session : boto3.Session(), optional
-        Boto3 Session. The default boto3 Session will be used if boto3_session receive None.
-
-    Returns
-    -------
-    None
-        None.
-
-    Examples
-    --------
-    Deleting a table
-
-    >>> import awswrangler as wr
-    >>> arn = wr.timestream.delete_table("MyDatabase", "MyTable")
-
-    """
-    _logger.info("Deleting Timestream table %s in database %s", table, database)
-    client = _utils.client(service_name="timestream-write", session=boto3_session)
-    client.delete_table(DatabaseName=database, TableName=table)
-
-
-def list_databases(
-    boto3_session: Optional[boto3.Session] = None,
-) -> List[str]:
-    """
-    List all databases in timestream.
-
-    Parameters
-    ----------
-    boto3_session : boto3.Session(), optional
-        Boto3 Session. The default boto3 Session will be used if boto3_session receive None.
-
-    Returns
-    -------
-    List[str]
-        a list of available timestream databases.
-
-    Examples
-    --------
-    Querying the list of all available databases
-
-    >>> import awswrangler as wr
-    >>> wr.timestream.list_databases()
-    ... ["database1", "database2"]
-
-
-    """
-    client = _utils.client(service_name="timestream-write", session=boto3_session)
-
-    response = client.list_databases()
-    dbs: List[str] = [db["DatabaseName"] for db in response["Databases"]]
-    while "NextToken" in response:
-        response = client.list_databases(NextToken=response["NextToken"])
-        dbs += [db["DatabaseName"] for db in response["Databases"]]
-
-    return dbs
-
-
-def list_tables(database: Optional[str] = None, boto3_session: Optional[boto3.Session] = None) -> List[str]:
-    """
-    List tables in timestream.
-
-    Parameters
-    ----------
-    database: str
-        Database name. If None, all tables in Timestream will be returned. Otherwise, only the tables inside the
-        given database are returned.
-    boto3_session : boto3.Session(), optional
-        Boto3 Session. The default boto3 Session will be used if boto3_session receive None.
-
-    Returns
-    -------
-    List[str]
-        A list of table names.
-
-    Examples
-    --------
-    Listing all tables in timestream across databases
-
-    >>> import awswrangler as wr
-    >>> wr.timestream.list_tables()
-    ... ["table1", "table2"]
-
-    Listing all tables in timestream in a specific database
-
-    >>> import awswrangler as wr
-    >>> wr.timestream.list_tables(DatabaseName="database1")
-    ... ["table1"]
-
-    """
-    client = _utils.client(service_name="timestream-write", session=boto3_session)
-    args = {} if database is None else {"DatabaseName": database}
-    response = client.list_tables(**args)  # type: ignore[arg-type]
-    tables: List[str] = [tbl["TableName"] for tbl in response["Tables"]]
-    while "nextToken" in response:
-        response = client.list_tables(**args, NextToken=response["NextToken"])  # type: ignore[arg-type]
-        tables += [tbl["TableName"] for tbl in response["Tables"]]
-
-    return tables
